+++ controller/src/main/java/com/linbit/linstor/dbdrivers/AbsDatabaseDriver.java
@@ -39,7 +39,7 @@ import java.util.function.Function;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-public abstract class AbsDatabaseDriver<DATA, INIT_MAPS, LOAD_ALL>
+public abstract class AbsDatabaseDriver<DATA extends Comparable<? super DATA>, INIT_MAPS, LOAD_ALL>
     implements GenericDatabaseDriver<DATA>, ControllerDatabaseDriver<DATA, INIT_MAPS, LOAD_ALL>
 {
     protected static final String MSG_DO_NOT_LOG = "do not log";