+++ controller/src/main/java/com/linbit/linstor/dbdrivers/k8s/crd/K8sCrdEngine.java
@@ -218,7 +218,7 @@ public class K8sCrdEngine implements DbEngine
     }
 
     @Override
-    public <DATA, INIT_MAPS, LOAD_ALL> Map<DATA, INIT_MAPS> loadAll(
+    public <DATA extends Comparable<? super DATA>, INIT_MAPS, LOAD_ALL> Map<DATA, INIT_MAPS> loadAll(
         DatabaseTable table,
         LOAD_ALL parents,
         DataLoader<DATA, INIT_MAPS, LOAD_ALL> dataLoader