+++ server/src/main/java/com/linbit/linstor/security/AccessControlEntry.java
@@ -1,11 +1,13 @@
 package com.linbit.linstor.security;
 
+import java.util.Objects;
+
 /**
  * Access control entry for access control lists
  *
  * @author Robert Altnoeder &lt;robert.altnoeder@linbit.com&gt;
  */
-public final class AccessControlEntry
+public final class AccessControlEntry implements Comparable<AccessControlEntry>
 {
     public final String objPath;
     public final Role subjectRole;
@@ -16,5 +18,37 @@ public final class AccessControlEntry
         objPath = objPathRef;
         subjectRole = subjRoleRef;
         access = accRef;
+    }
+
+    @Override
+    public int compareTo(AccessControlEntry oRef)
+    {
+        int cmp = objPath.compareTo(oRef.objPath);
+        if (cmp == 0)
+        {
+            cmp = subjectRole.compareTo(oRef.subjectRole);
+        }
+        return cmp;
+    }
+
+    @Override
+    public int hashCode()
+    {
+        return Objects.hash(objPath, subjectRole);
+    }
+
+    @Override
+    public boolean equals(Object obj)
+    {
+        if (this == obj)
+        {
+            return true;
+        }
+        if (!(obj instanceof AccessControlEntry))
+        {
+            return false;
+        }
+        AccessControlEntry other = (AccessControlEntry) obj;
+        return Objects.equals(objPath, other.objPath) && Objects.equals(subjectRole, other.subjectRole);
     }
 }