+++ server/src/main/java/com/linbit/linstor/storage/data/adapter/drbd/DrbdRscDfnData.java
@@ -34,7 +34,7 @@ import java.util.Objects;
 
 public class DrbdRscDfnData<RSC extends AbsResource<RSC>>
     extends BaseTransactionObject
-    implements DrbdRscDfnObject, Comparable<DrbdRscDfnData<RSC>>
+    implements DrbdRscDfnObject
 {
     public static final int SNAPSHOT_TCP_PORT = -1;
 
@@ -313,37 +313,5 @@ public class DrbdRscDfnData<RSC extends AbsResource<RSC>>
                 Objects.equals(rscName, other.rscName) && Objects.equals(snapName, other.snapName);
         }
         return ret;
-    }
-
-    @Override
-    public int compareTo(DrbdRscDfnData<RSC> oRef)
-    {
-        int cmp = rscName.compareTo(oRef.rscName);
-        if (cmp == 0)
-        {
-            if (snapName != null)
-            {
-                if (oRef.snapName == null)
-                {
-                    cmp = -1;
-                }
-                else
-                {
-                    cmp = snapName.compareTo(oRef.snapName);
-                }
-            }
-            else
-            {
-                if (oRef.snapName != null)
-                {
-                    cmp = 1;
-                }
-            }
-            if (cmp == 0)
-            {
-                cmp = resourceNameSuffix.compareTo(oRef.resourceNameSuffix);
-            }
-        }
-        return cmp;
     }
 }