+++ server/src/main/java/com/linbit/linstor/storage/interfaces/categories/resource/VlmDfnLayerObject.java
@@ -1,5 +1,6 @@
 package com.linbit.linstor.storage.interfaces.categories.resource;
 
+import com.linbit.GenericName;
 import com.linbit.linstor.api.interfaces.VlmDfnLayerDataApi;
 import com.linbit.linstor.core.identifier.ResourceName;
 import com.linbit.linstor.core.identifier.SnapshotName;
@@ -7,11 +8,12 @@ import com.linbit.linstor.core.identifier.VolumeNumber;
 import com.linbit.linstor.dbdrivers.DatabaseException;
 import com.linbit.linstor.security.AccessContext;
 import com.linbit.linstor.storage.interfaces.categories.LayerObject;
+import com.linbit.utils.StringUtils;
 
 /**
  * @author Gabor Hernadi &lt;gabor.hernadi@linbit.com&gt;
  */
-public interface VlmDfnLayerObject extends LayerObject
+public interface VlmDfnLayerObject extends LayerObject, Comparable<VlmDfnLayerObject>
 {
     String getRscNameSuffix();
 
@@ -24,4 +26,23 @@ public interface VlmDfnLayerObject extends LayerObject
     void delete() throws DatabaseException;
 
     VlmDfnLayerDataApi getApiData(AccessContext accCtxRef);
+
+    @Override
+    default int compareTo(VlmDfnLayerObject oRef)
+    {
+        int cmp = getResourceName().compareTo(oRef.getResourceName());
+        if (cmp == 0)
+        {
+            cmp = StringUtils.compareToNullable(getRscNameSuffix(), oRef.getRscNameSuffix());
+            if (cmp == 0)
+            {
+                cmp = GenericName.compareToNullable(getSnapshotName(), oRef.getSnapshotName());
+                if (cmp == 0)
+                {
+                    cmp = getVolumeNumber().compareTo(oRef.getVolumeNumber());
+                }
+            }
+        }
+        return cmp;
+    }
 }