+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Force all database DATA objects to be comparable
+
 ## [1.24.0-rc.1] - 2023-07-20
 
 ### Added